Suunto implementation for CloudService
Added new implementation of CloudService for Suunto, allowing the
retrieval of activity details from their Suunto App platform (not the
legacy Movescount platform). Only OAuth, searching for activities, and
downloading activities are supported. Suunto does not have an API for
uploading activities, and routes/step activities are not suppported by
this implementation.

New variables that are required to be set are listed in the
gcconfig.pri.in template file.

To go along with this, the .fit file parsing in FitRideFile now supports
detecting Suunto as a manufacturer and identifying watch models from
devices.